Move defaults to tools/pytest_fixtures/webdriver_defaults.py
Issue: #59157
